chore: Move trigram index into tracker [DHIS2-17714]
